Renamed @homomorphism to @hom.  Deleted substitute_symbols_homomorphism, and now just call substitute_symbols, which I've moved to Syntax.
